add electron-manager
- command run now returns the childprocess which contains electron with
  IPC
- implement arguments can now passed through run command
