+++ src/lib/main.ts
@@ -4,16 +4,16 @@ import * as vscode from "vscode";
 import { ELECTRON_INSTALL_PATH, ELECTRON_VERSION, OUTPUT_CHANNEL } from "./api";
 
 import { ElectronInstaller } from "./installer";
-import { ElectronStarter } from "./starter";
+import { ElectronManager } from "./manager";
 
 export function activate(context: vscode.ExtensionContext) {
 
-    container.register(ELECTRON_INSTALL_PATH, {useValue: "./electron"});
-    container.register(ELECTRON_VERSION, {useValue: "11.1.0"});
+    container.register(ELECTRON_INSTALL_PATH, {useValue: "./electron"})
+    container.register(ELECTRON_VERSION, {useValue: "11.1.0"})
     container.register(OUTPUT_CHANNEL, {useValue: vscode.window.createOutputChannel('QMasters:Electron')})
 
     const installer = container.resolve(ElectronInstaller)
-    const electronStarter = container.resolve(ElectronStarter)
+    const electronManager = container.resolve(ElectronManager)
 
     let disposables = [
         /**
@@ -26,9 +26,7 @@ export function activate(context: vscode.ExtensionContext) {
          * run electron
          *
          */
-        vscode.commands.registerCommand('qmasters:electron.run', async (file: string) => {
-            electronStarter.run(file)
-        })
+        vscode.commands.registerCommand('qmasters:electron.run', async (file: string, ...args) => electronManager.run(file, args))
     ]
 
     context.subscriptions.push(...disposables)
