+++ src/package.json
@@ -54,7 +54,7 @@
     "ts-loader": "8.0.17",
     "tsyringe": "4.4.0",
     "typescript": "4.1.5",
-    "vsce": "^1.85.0",
+    "vsce": "1.85.0",
     "webpack": "5.21.2",
     "webpack-cli": "4.5.0"
   },
